refactor: rename Outcome to OutcomeTag
