+++ packages/app-builder/src/components/Decisions/index.ts
@@ -3,5 +3,5 @@ export * from './DecisionRightPanel';
 export * from './decisions-i18n';
 export * from './DecisionsList';
 export * from './Filters';
-export * from './Outcome';
+export * from './OutcomeTag';
 export * from './RulesDetail';
